Organize space combat autoresolution
