+++ Assets/Scripts/Game/Combat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ef0e9211151e4b6f96418fe6c7caff4f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
