+++ Assets/Scripts/Game/Combat/SpaceCombatAutoResolution.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9bbb948e69f94107b7711b7220ddac1b
