+++ Assets/Tests/EditMode/GameTests/Game/Combat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 42d638b0061a41b3a3bca6bb7065ba60
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
